+++ src/array.js
@@ -1,7 +1,7 @@
 import _ from 'lodash/fp'
 import { callOrReturn } from './function'
 import { insertAtIndex } from './collection'
-import { reduceIndexed } from './conversion'
+import { reduceIndexed, mapIndexed } from './conversion'
 
 // TODO: Move to proper files and expose
 let callUnless = (check) => (failFn) => (fn) => (x, y) =>
@@ -16,6 +16,7 @@ let last = _.takeRight(1)
  * Joins an array after compacting. Note that due to the underlying behavior of `_.curry` no default `join` value is supported -- you must pass in some string with which to perform the join.
  *
  * @signature joinString -> [string1, string2, ...stringN] -> string1 + joinString + string2 + joinString ... + stringN
+ * @typescript {(join: string, x: any[]) => string}
  */
 export let compactJoin = _.curry((join, x) => _.compact(x).join(join))
 
@@ -23,6 +24,7 @@ export let compactJoin = _.curry((join, x) => _.compact(x).join(join))
  * Compacts and joins an array with `.`
  *
  * @signature [string1, string2, ...stringN] -> string1 + '.' + string2 + '.' ... + stringN
+ * @typescript {(arr: any[]) => string}
  */
 export let dotJoin = compactJoin('.')
 
@@ -31,7 +33,7 @@ export let dotJoin = compactJoin('.')
  *
  * @signature filterFunction -> [string1, string2, ...stringN] -> string1 + '.' + string2 + '.' ... + stringN
  */
-export let dotJoinWith = (fn) => (x) => _.filter(fn, x).join('.')
+export let dotJoinWith = (fn) => (x) => _.flow(_.filter(fn), _.join('.'))(x)
 
 /**
  * Returns an array of elements that are repeated in the array.
@@ -90,6 +92,15 @@ export let mergeRanges = _.flow(
 )
 
 /**
+ * Determines if an array is a subset of another array.
+ *
+ * @signature ([a], [a]) -> boolean
+ */
+export let isSubset = _.curry(
+  (array1, array2) => _.difference(array1, array2).length === 0
+)
+
+/**
  * Creates a function that takes an element of the original array as argument and returns the next element in the array (with wrapping). Note that (1) This will return the first element of the array for any argument not in the array and (2) due to the behavior of `_.curry` the created function will return a function equivalent to itself if called with no argument.
  *
  * @signature [a, b...] -> a -> b
@@ -102,13 +113,14 @@ export let cycle = _.curry((a, n) => a[(a.indexOf(n) + 1) % a.length])
  * @signature (k, v, [a]) -> { k(a): v(a) }
  */
 export let arrayToObject = _.curry((k, v, a) =>
-  _.flow(_.keyBy(k), _.mapValues(v))(a)
+  _.zipObject(mapIndexed(k, a), mapIndexed(v, a))
 )
 
 /**
  * Converts and array of keys to an object using a predicate
  *
  * @signature (v, [a]) => { a: v(a) }
+ * @typescript <T>(fn: (k: string) => T, keys: string[]): { [K in typeof keys[number]]: T } // TS not enforcing the keys :(
  */
 export let keysToObject = arrayToObject((x) => x)
 
@@ -177,6 +189,19 @@ export let chunkBy = _.curry((f, array) =>
         [[_.head(array)]],
         _.tail(array)
       )
+)
+
+/**
+ * `chunkBy` when the returned value of an iteratee changes
+ *
+ * @signature f -> [] -> [[], ...]
+ * @since 1.75.0
+ */
+export let chunkByValue = _.curry((f, array) =>
+  chunkBy(
+    (group, fn) => _.isEqual(_.iteratee(f)(_.last(group)), _.iteratee(f)(fn)),
+    array
+  )
 )
 
 /**